+++ lib/astarte_rpc/protocol/proto/realm_management/get_detailed_interfaces_list_reply.pb.ex
@@ -1,0 +1,7 @@
+defmodule Astarte.RPC.Protocol.RealmManagement.GetDetailedInterfacesListReply do
+  @moduledoc false
+
+  use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
+
+  field :interface_json, 1, repeated: true, type: :string, json_name: "interfaceJson"
+end