+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   number of registered devices per realm.
 - Add `GetDatastreamMaximumStorageRetention[Reply]` to retrieve the maximum
   retention value for data in a realm.
+- Add `GetDetailedInterfacesList[Reply]` to retrieve the detailed
+  interfaces list in a realm.
 
 ### Changed
 - Bump Elixir to 1.15.7.